requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.1
Django==4.1.5
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-filter==22.1
django-stdimage==6.0.1
gunicorn==20.1.0
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.4.0
pycodestyle==2.10.0
PyMySQL==1.0.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
whitenoise==6.3.0
